Fixed bug in GenericRepository, set up automapper to mapp models from other assemblies.

+++ Source/EverlastingStudent/EverlastingStudent.DataTransferObjects/GetHomeworksDto.cs
@@ -13,9 +13,9 @@
 
         public int EnergyCost { get; set; }
 
-        public int KnowledgeGain { get; set; }
+        public double KnowledgeGain { get; set; }
 
-        public int ExperienceGain { get; set; }
+        public double ExperienceGain { get; set; }
 
         public int SolveDurabationInMinutes { get; set; }
